slice-5d.1: introduce ShareIfcManager + ShareIfcLoader wrappers
Build two new classes that wrap Conway's IfcAPI directly, replacing
the live consumer paths through `viewer.IFC.loader` (web-ifc-three's
`IFCLoader`) and `viewer.IFC.loader.ifcManager` (web-ifc-three's
`IFCManager`):

- `src/viewer/ifc/ShareIfcManager.js` — IFC manager surface. Owns:
  `ifcAPI`, `state.models` (legacy bookkeeping), `parser` (stub `{}`
  for `IfcViewsManager` to mutate harmlessly), `getExpressId`,
  `getSpatialStructure`, `getItemProperties`, `getPropertySets`,
  `getIfcType` (type-code variant), `idsByType`. Methods the fork's
  IFCManager exposed but live code no longer needs
  (`applyWebIfcConfig`, `setupCoordinationMatrix`, `loadIfc`, …) are
  intentionally absent — if anything tries to reach for them the
  failure is loud.

- `src/viewer/ifc/ShareIfcLoader.js` — IFC loader. Owns the
  Conway-direct `parse(buffer, onProgress, onError)` entry point that
  `Loader.js#readModel` invokes. Holds a `ShareIfcManager` accessible
  as `loader.ifcManager`. Mirrors `ifcLastError` writes to
  `viewer.IFC.ifcLastError` (the canonical slot consumers read from).

- `src/viewer/ifc/ifcItemsMapParity.js` — `runIfcItemsMapParityCheck`
  extracted from `Loader.js` so it lives next to the IFC parse path
  it diagnoses. Pure refactor; no behavior change.

ShareViewer's constructor now installs the loader after composing
the fork: `this.IFC.loader = new ShareIfcLoader({ifcAPI, ifc})`. The
Conway IfcAPI handle is still sourced from
`this._fork.IFC.loader.ifcManager.ifcAPI` during this transitional
slice; later slices instantiate Conway directly via
`@bldrs-ai/conway-web-ifc-adapter`.

`Loader.js` shrinks:
- `newIfcLoader(viewer)` deleted — the hot-patch logic moved into
  `ShareIfcLoader.parse` verbatim. The `case 'ifc'` arm now reads
  `loader = viewer.IFC.loader` (= the ShareIfcLoader instance).
- `runIfcItemsMapParityCheck` moved to its own file.
- The unused `compareItemsMaps` / `formatComparison` /
  `itemsMapFromFlatMeshes` / `itemsMapFromPerVertexAttribute` /
  `buildConwayIfcModel` / `decorateConwayDirectIfcModel` /
  `parseIfcWithConway` imports go with them.

Test updates: `Loader.test.js` and `Loader.cover.test.js` build their
own viewer stubs (not via `__mocks__/web-ifc-viewer.js`), so they
install a real `ShareIfcLoader` against their existing `ifcAPI` mock.
The previous `loader.parse: jest.fn()` stubs were vestigial once
`newIfcLoader` mutated the parse onto `viewer.IFC` — they're left in
place but unused.

Net diff: 4 modified files (-256 lines from Loader.js including
moved code, +51 elsewhere), 3 new files. 1707/1707 Jest passing.
