+++ src/loader/Loader.cover.test.js
@@ -7,6 +7,7 @@
 
 import axios from 'axios'
 import {downloadToOPFS, downloadModel, getModelFromOPFS} from '../OPFS/utils'
+import ShareIfcLoader from '../viewer/ifc/ShareIfcLoader'
 import {constructUploadedBlobPath, load, NotFoundError} from './Loader'
 import {dereferenceAndProxyDownloadContents} from './urls'
 
@@ -429,6 +430,8 @@ describe('load() error/edge paths with OPFS enabled', () => {
         ifcLastError: null,
         addIfcModel: jest.fn(),
         loader: {
+          // Replaced below with a real ShareIfcLoader so the OOM
+          // throw from `OpenModel` flows through our parse path.
           parse: jest.fn().mockRejectedValue(oomErr),
           ifcManager: {
             state: {models: []},
@@ -460,6 +463,14 @@ describe('load() error/edge paths with OPFS enabled', () => {
         },
       },
     }
+    // Slice 5d.1: install ShareIfcLoader. The OOM-throwing OpenModel
+    // is on `ifcViewer.IFC.loader.ifcManager.ifcAPI` (set above);
+    // ShareIfcLoader.parse calls into it and surfaces the OOM the
+    // same way `newIfcLoader.parse` did pre-5d.1.
+    ifcViewer.IFC.loader = new ShareIfcLoader({
+      ifcAPI: ifcViewer.IFC.loader.ifcManager.ifcAPI,
+      ifc: ifcViewer.IFC,
+    })
 
     dereferenceAndProxyDownloadContents.mockResolvedValue([
       'https://example.com/model.ifc',
@@ -517,6 +528,14 @@ describe('load() error/edge paths with OPFS enabled', () => {
       },
     }
     const primedViewer = {IFC: ifcLoaderBase}
+    // Slice 5d.1: install ShareIfcLoader. The "model already present"
+    // guard lives inside ShareIfcLoader.parse — it inspects
+    // `context.items.ifcModels.length`, which we've pre-populated
+    // above so the guard fires immediately.
+    primedViewer.IFC.loader = new ShareIfcLoader({
+      ifcAPI: primedViewer.IFC.loader.ifcManager.ifcAPI,
+      ifc: primedViewer.IFC,
+    })
 
     dereferenceAndProxyDownloadContents.mockResolvedValue([
       'https://example.com/model.ifc',
