+++ src/loader/Loader.js
@@ -26,14 +26,6 @@ import {updateRecentFileLastModified} from '../connections/persistence'
 import {testUuid} from '../utils/strings'
 import {decorateShareModel, inferModelCapabilities} from '../viewer/ShareModel'
 import {attachElementSubsets, attachInstanceMapSubsets, summariseElementIdAttribute} from '../viewer/three/elementSubsets'
-import {
-  compareItemsMaps,
-  formatComparison,
-  itemsMapFromFlatMeshes,
-  itemsMapFromPerVertexAttribute,
-} from '../viewer/ifc/IfcItemsMap'
-import {buildConwayIfcModel} from '../viewer/ifc/buildConwayIfcModel'
-import {decorateConwayDirectIfcModel, parseIfcWithConway} from '../viewer/ifc/conwayDirectIfcLoader'
 import {instanceMapFromGeometry, instanceMapFromTriangleIds} from '../viewer/ifc/IfcInstanceMap'
 import {dereferenceAndProxyDownloadContents} from './urls'
 import BLDLoader from './BLDLoader'
@@ -1225,7 +1217,11 @@ async function findLoader(pathname, viewer) {
     case 'step':
     case 'stp':
     case 'ifc': {
-      loader = newIfcLoader(viewer)
+      // ShareIfcLoader (`src/viewer/ifc/ShareIfcLoader.js`) is wired
+      // up as `viewer.IFC.loader` at viewer construction (see
+      // `src/viewer/ShareViewer.js`). It owns the Conway-direct parse
+      // entry point that `readModel` invokes below.
+      loader = viewer.IFC.loader
       isLoaderAsync = true
       // TODO(pablo): true should work but currently causes IFCLoader to fail
       isFormatText = false
@@ -1331,304 +1327,6 @@ function newGltfLoader() {
     loader.setMeshoptDecoder(MeshoptDecoder)
   }
   return loader
-}
-
-
-/**
- * Sets up the IFCLoader to use the wasm module and move the model to
- * the origin on load.
- *
- * @param {object} viewer
- * @return {object} Loader with parse function
- */
-function newIfcLoader(viewer) {
-  const loader = viewer.IFC
-  // Track last IFC parse error (especially when parse returns null)
-  loader.ifcLastError = null
-  // Slice 5b of design/new/viewer-replacement.md Phase 5: IFC parse
-  // goes through Conway directly. Wit-three's `IFCLoader.parse`
-  // (which used to drive Conway under the hood and emit a
-  // throw-away geometry we'd rebuild via the Conway-direct
-  // assembler) is no longer called — `parseIfcWithConway` calls
-  // `ifcAPI.OpenModel` + `StreamAllMeshes` ourselves, then
-  // `buildConwayIfcModel` produces the renderable Mesh in one pass.
-  // The wit-three `ifcManager` is still around for `addIfcModel` +
-  // `context.items` access; Slice 5c drops that too.
-  loader.parse = async function(
-    buffer,
-    onProgress,
-    onError,
-  ) {
-    if (this.context.items.ifcModels.length !== 0) {
-      throw new Error('Model cannot be loaded.  A model is already present')
-    }
-    try {
-      if (onProgress) {
-        onProgress('Parsing model geometry...')
-      }
-      const ifcAPI = this.loader.ifcManager.ifcAPI
-      const {modelID, captured} = await parseIfcWithConway(buffer, ifcAPI)
-
-      if (onProgress) {
-        onProgress('Building model...')
-      }
-      const {mesh: ifcModel, stats: buildStats} = buildConwayIfcModel(
-        captured, ifcAPI, modelID)
-      const scene = typeof this.context?.getScene === 'function' ?
-        this.context.getScene() : null
-      decorateConwayDirectIfcModel(ifcModel, ifcAPI, modelID, {scene})
-
-      this.addIfcModel(ifcModel)
-
-      if (onProgress) {
-        onProgress('Setting up coordinate system...')
-      }
-      // eslint-disable-next-line new-cap
-      const matrixArr = await ifcAPI.GetCoordinationMatrix(modelID)
-      // Apply the coordination matrix to the model directly. Wit-three's
-      // `setupCoordinationMatrix` set this on the model + told the
-      // IFCParser to re-apply on every subsequent mesh; with Conway-
-      // direct there's no IFCParser, so a one-shot apply is enough
-      // and the matrix can be stamped onto the Mesh's transform.
-      //
-      // Optional-chained for test resilience — `jest.mock('three')`
-      // mocked Mesh instances don't have a real `Matrix4` for
-      // `ifcModel.matrix`. Real three.js Mesh always does.
-      if (ifcModel.matrix && typeof ifcModel.matrix.fromArray === 'function') {
-        ifcModel.matrix.fromArray(matrixArr)
-        ifcModel.matrixAutoUpdate = false
-      }
-
-      if (onProgress) {
-        onProgress('Fitting model to frame...')
-      }
-      this.context.fitToFrame()
-
-      if (onProgress) {
-        onProgress('Gathering model statistics...')
-      }
-      const statsApi = ifcAPI.getStatistics(modelID)
-      ifcModel.name = statsApi.projectName ?? undefined
-      const loadStats = {
-        loaderVersion: ifcAPI.getConwayVersion(),
-        geometryMemory: statsApi.getGeometryMemory(),
-        geometryTime: statsApi.getGeometryTime(),
-        ifcVersion: statsApi.getVersion(),
-        loadStatus: statsApi.getLoadStatus(),
-        originatingSystem: statsApi.getOriginatingSystem(),
-        preprocessorVersion: statsApi.getPreprocessorVersion(),
-        parseTime: statsApi.getParseTime(),
-        totalTime: statsApi.getTotalTime(),
-      }
-      ifcModel.loadStats = loadStats
-
-      if (onProgress) {
-        onProgress('Model loaded successfully!')
-      }
-
-      // Parallel-run the new IfcItemsMap populators against the live
-      // model and log the diff. Diagnostic only — no behavior change.
-      // Toggle via `?feature=ifcItemsMapParity`. See
-      // design/new/viewer-replacement.md §3b.ii for the per-vertex-vs-
-      // per-instance story this check exposes.
-      if (isFeatureEnabled('ifcItemsMapParity')) {
-        runIfcItemsMapParityCheck(ifcAPI, ifcModel, captured)
-      }
-      // Always-on integration-boundary log. `conwayDirect.spec.ts`
-      // (and the deploy-preview smoke checks) gate on `[conwayDirect]
-      // parsed modelID=…` firing — it's the single observable signal
-      // that the Conway-direct parse + assembly path completed
-      // successfully on a real model. Kept at info level (not gated
-      // on glbVerbose) so the signal is visible in production logs
-      // without the user opting in to the verbose channel.
-      // eslint-disable-next-line no-console
-      console.info(
-        `[conwayDirect] parsed modelID=${modelID} — ` +
-        `vertices=${buildStats.vertexCount} triangles=${buildStats.triangleCount} ` +
-        `instances=${buildStats.instanceCount} parents=${buildStats.parentCount} ` +
-        `materials=${buildStats.materialCount} ` +
-        `skippedFlatMeshes=${buildStats.skippedFlatMeshes} ` +
-        `skippedPlaced=${buildStats.skippedPlacedGeometries}`)
-
-      return ifcModel
-    } catch (err) {
-      loader.ifcLastError = err
-      // Rethrow OOM so callers can present a tailored UX message.
-      if (isOutOfMemoryError(err)) {
-        err.isOutOfMemory = true // tag for convenience
-        throw err
-      }
-      console.error(err)
-      if (onError) {
-        onError(err)
-      }
-      return null
-    }
-  }
-  return loader
-}
-
-
-/**
- * Build the new IfcItemsMap two ways (per-vertex attribute and
- * Conway-direct stream) for a freshly-parsed IFC model, then log a
- * one-line comparison. Diagnostic only — pure observation; the
- * caller drops both maps on the floor.
- *
- * What we're watching for: an `onlyInB` count greater than zero
- * (Conway-direct sees instance IDs the per-vertex attribute
- * collapses). That's the IfcMappedItem case — design/new/viewer-replacement.md
- * §3b.ii. A clean model with no mapped items prints `both=N
- * onlyA=0 onlyB=0` and tells us the per-vertex path is already
- * per-instance for this file.
- *
- * `triCountDeltas > 0` would be a stronger signal: the two
- * populators agree on the ID set but disagree on triangle counts
- * per ID. That implies an emission-order mismatch and should be
- * investigated before promoting Conway-direct.
- *
- * Safe to fail silently — if Conway's API surface differs from what
- * we expect (e.g., a future adapter version drops `StreamAllMeshes`),
- * the parity check warns once and returns; the load completes
- * normally.
- *
- * @param {object} ifcAPI Conway-compatible IfcAPI
- * @param {object} ifcModel freshly-parsed web-ifc-three Mesh
- */
-/**
- * Build the new IfcItemsMap two ways (per-vertex attribute and
- * captured FlatMesh stream) for a freshly-parsed IFC model, then log
- * a one-line comparison. Diagnostic only — pure observation; the
- * caller drops both maps on the floor.
- *
- * @param {object} ifcAPI Conway-compatible IfcAPI
- * @param {object} ifcModel freshly-parsed web-ifc-three Mesh
- * @param {Array} capturedFlatMeshes FlatMeshes captured during the
- *   live parse via installFlatMeshCapture
- */
-function runIfcItemsMapParityCheck(ifcAPI, ifcModel, capturedFlatMeshes) {
-  try {
-    const modelID = ifcModel.modelID
-    const geom = ifcModel.geometry
-    const perVertex = itemsMapFromPerVertexAttribute(geom)
-    if (!perVertex) {
-      console.warn(
-        '[ifcItemsMapParity] per-vertex populator returned null ' +
-        '(no expressID attribute or no index) — skipping parity check')
-      return
-    }
-    if (!Array.isArray(capturedFlatMeshes) || capturedFlatMeshes.length === 0) {
-      console.warn(
-        '[ifcItemsMapParity] no FlatMesh capture from parse — ' +
-        'skipping parity check (web-ifc-three may not have called ' +
-        'StreamAllMeshes for this load)')
-      return
-    }
-    const conway = itemsMapFromFlatMeshes(
-      capturedFlatMeshes, ifcAPI, modelID, {geometry: geom})
-    const cmp = compareItemsMaps(perVertex, conway)
-    // Conway-level shape: how many PlacedGeometries fit under how
-    // many FlatMeshes? `placedCount > flatMeshCount` means some
-    // FlatMeshes have multiple PlacedGeometries. Two distinct
-    // sub-cases hide there:
-    //   (i)  IfcMappedItem-style instances: many PlacedGeometries
-    //        SHARE one `geometryExpressID` and differ only in their
-    //        flatTransformation. Per-instance picking is the right
-    //        answer for these.
-    //   (ii) Compound representation: one IFC element's geometry is
-    //        built from N distinct geometric primitives, each with
-    //        its own `geometryExpressID`. Subcomponents are not
-    //        independently selectable in IFC semantics; per-instance
-    //        picking would be wrong here.
-    // Tracking shared vs unique geometryExpressIDs across the
-    // multi-placed FlatMeshes tells us the mix.
-    let flatMeshCount = 0
-    let placedCount = 0
-    let multiPlacedFlatMeshes = 0
-    let maxPlacedInOneFlatMesh = 0
-    // Case (i) and (ii) accounting across multi-placed FlatMeshes:
-    let sharedShapeInstances = 0 // PlacedGeometries that share a geomExpressID with a sibling
-    let uniqueShapeInstances = 0 // PlacedGeometries with a geomExpressID unique within the FlatMesh
-    let multiPlacedAllShared = 0 // FlatMeshes where every placedGeom shares one geomExpressID (pure case i)
-    let multiPlacedAllUnique = 0 // FlatMeshes where every placedGeom has a unique geomExpressID (pure case ii)
-    let multiPlacedMixed = 0 // FlatMeshes that are a mix
-    for (const fm of capturedFlatMeshes) {
-      flatMeshCount++
-      const g = fm?.geometries
-      const n = typeof g?.size === 'function' ? g.size() : (g?.length ?? 0)
-      placedCount += n
-      if (n > 1) {
-        multiPlacedFlatMeshes++
-        // Count occurrences of each geomExpressID inside this FlatMesh.
-        const counts = new Map()
-        for (let i = 0; i < n; i++) {
-          const placed = typeof g.get === 'function' ? g.get(i) : g[i]
-          const id = placed?.geometryExpressID
-          counts.set(id, (counts.get(id) ?? 0) + 1)
-        }
-        let sharedHere = 0
-        let uniqueHere = 0
-        for (const c of counts.values()) {
-          if (c > 1) {
-            sharedHere += c
-          } else {
-            uniqueHere += c
-          }
-        }
-        sharedShapeInstances += sharedHere
-        uniqueShapeInstances += uniqueHere
-        if (sharedHere === n) {
-          multiPlacedAllShared++
-        } else if (uniqueHere === n) {
-          multiPlacedAllUnique++
-        } else {
-          multiPlacedMixed++
-        }
-      }
-      if (n > maxPlacedInOneFlatMesh) {
-        maxPlacedInOneFlatMesh = n
-      }
-    }
-    console.warn(
-      `[ifcItemsMapParity] modelID=${modelID} ` +
-      `perVertexElements=${perVertex.elementCount} ` +
-      `conwayElements=${conway.elementCount} ` +
-      `perVertexTriangles=${perVertex.triangleCount} ` +
-      `conwayTriangles=${conway.triangleCount} ` +
-      `${formatComparison(cmp)}`)
-    console.warn(
-      `[ifcItemsMapParity] Conway emission: flatMeshes=${flatMeshCount} ` +
-      `placedGeometries=${placedCount} ` +
-      `multiPlacedFlatMeshes=${multiPlacedFlatMeshes} ` +
-      `maxPlacedInOneFlatMesh=${maxPlacedInOneFlatMesh}`)
-    console.warn(
-      `[ifcItemsMapParity] multi-placed breakdown: ` +
-      `allShared=${multiPlacedAllShared} ` +
-      `allUnique=${multiPlacedAllUnique} ` +
-      `mixed=${multiPlacedMixed} ` +
-      `sharedShapeInstances=${sharedShapeInstances} ` +
-      `uniqueShapeInstances=${uniqueShapeInstances}`)
-    // First few count deltas, for spot-checking emission-order issues.
-    if (cmp.triangleCountDeltas.length > 0) {
-      const head = cmp.triangleCountDeltas.slice(0, 5)
-        .map((d) => `${d.id}:${d.a}vs${d.b}`).join(' ')
-      console.warn(`[ifcItemsMapParity] sample triCountDeltas: ${head}`)
-    }
-    if (cmp.onlyInB > 0) {
-      console.warn(
-        `[ifcItemsMapParity] Conway sees ${cmp.onlyInB} additional ` +
-        'instance IDs — this is the IfcMappedItem per-instance ' +
-        'delta the per-vertex path collapses')
-    }
-    if (!cmp.agreeingTriangleCounts) {
-      console.warn(
-        `[ifcItemsMapParity] ${cmp.triangleCountDeltas.length} IDs ` +
-        'have triangle-count deltas — emission order may differ; ' +
-        'investigate before promoting Conway-direct')
-    }
-  } catch (e) {
-    console.warn('[ifcItemsMapParity] check failed:', e)
-  }
 }
