+++ src/loader/Loader.test.js
@@ -1,5 +1,6 @@
 import {Object3D, Mesh, BufferGeometry, Material, BufferAttribute} from 'three'
 import {GLTFLoader} from 'three/examples/jsm/loaders/GLTFLoader.js'
+import ShareIfcLoader from '../viewer/ifc/ShareIfcLoader'
 import {load, readModel} from './Loader'
 
 
@@ -78,6 +79,19 @@ describe('Loader', () => {
         },
       },
     }
+    // Slice 5d.1: production wires `viewer.IFC.loader` to a
+    // ShareIfcLoader during ShareViewer construction. This test
+    // builds its own viewer stub (not via __mocks__/web-ifc-viewer),
+    // so we install a real ShareIfcLoader here so `loader.parse`
+    // runs the Conway-direct flow against the mock ifcAPI. The
+    // previous `loader.parse: jest.fn().mockResolvedValue(...)` stub
+    // was overridden by `newIfcLoader` pre-5d.1 — it's vestigial
+    // now and kept only for any test that still grabs `parse` off
+    // the inner ifcManager via legacy reads.
+    mockViewer.IFC.loader = new ShareIfcLoader({
+      ifcAPI: mockViewer.IFC.loader.ifcManager.ifcAPI,
+      ifc: mockViewer.IFC,
+    })
   })
   afterEach(() => mathRandomSpy.mockRestore())
